+++ src/components/Player/Player.tsx
@@ -1,7 +1,13 @@
+import { FC } from 'react';
+import Track from '../../models/track.model';
 import './Player.css';
 
-const Player = () => {
-  return <div className='player'>Player</div>;
+interface PlayerProps {
+  track?: Track;
+}
+
+const Player: FC<PlayerProps> = ({ track }) => {
+  return <div className="player">Player</div>;
 };
 
 export default Player;