+++ src/constants/playlist.constants.ts
@@ -2,3 +2,5 @@ export const DEFAULT_IMG_URL =
   'https://images.pexels.com/photos/14925306/pexels-photo-14925306.jpeg';
 
 export const PLAY_VISIBLE_SCROLL_TOP = 189;
+
+export const FAVORITES_PLAYLIST_ID = 'favorites';