+++ src/containers/FavoritesContainer/useFavorites.tsx
@@ -1,48 +1,57 @@
-import { useMemo, useState } from 'react';
+import { useContext, useState } from 'react';
+import favoritesImg from '../../assets/like.png';
+import { FAVORITES_PLAYLIST_ID } from '../../constants/playlist.constants';
+import PlayerContext from '../../contexts/PlayerContext';
 import { PlaylistType } from '../../enums/playlistType.enum';
 import Playlist from '../../models/playlist.model';
-import PlaylistTrack from '../../models/playlistTrack.model';
-import {
-  addFavorite,
-  getFavorites,
-  removeFavorite,
-} from '../../utils/storage.utils';
-import favoritesImg from '../../assets/like.png';
+import { getFavorites, removeFavorite } from '../../utils/storage.utils';
 
 const useFavorites = () => {
-  //#region [states]
-  const [favoritesObj, setFavoritesObj] = useState<
-    Record<string, PlaylistTrack>
-  >(getFavorites());
+  //#region [contexts]
+  const {
+    currentPlaylist,
+    currentTrackIndex,
+    setCurrentTrackIndex,
+    setCurrentPlaylist,
+  } = useContext(PlayerContext);
   //#endregion
 
-  //#region [memos]
-  const favoritesPlaylist = useMemo(() => {
-    return {
-      id: 'favorites',
-      tracks: Object.values(favoritesObj),
-      name: 'Titres favoris',
-      url: 'collections/favorites',
-      users: ['Nicolas Bodin'],
-      description: 'Tous les titres que vous adorez !',
-      imgUrl: favoritesImg,
-      type: PlaylistType.COLLECTION,
-    } as Playlist;
-  }, [favoritesObj]);
+  //#region [states]
+  const [favoritesPlaylist, setFavoritesPlaylist] = useState<Playlist>({
+    id: FAVORITES_PLAYLIST_ID,
+    tracks: Object.values(getFavorites()),
+    name: 'Titres favoris',
+    url: `collection/${FAVORITES_PLAYLIST_ID}`,
+    users: ['Nicolas Bodin'],
+    description: 'Tous les titres que vous adorez !',
+    imgUrl: favoritesImg,
+    type: PlaylistType.COLLECTION,
+  });
   //#endregion
 
   //#region [handle methods]
   const handleLikeClick = (trackIndex: number) => {
-    const { id, favorite } = favoritesPlaylist.tracks[trackIndex].track;
-    if (favorite) {
-      removeFavorite(id);
-    } else {
-      addFavorite(favoritesPlaylist.tracks[trackIndex].track);
+    // on récupère l'identifiant du favori qu'on vient d'enlever
+    const { id } = favoritesPlaylist.tracks[trackIndex].track;
+
+    // on met à jour le state en supprimant le favori
+    const playlistCopy = { ...favoritesPlaylist };
+    playlistCopy.tracks = playlistCopy.tracks.filter(
+      (playlistTrack) => playlistTrack.track.id !== id,
+    );
+    setFavoritesPlaylist(playlistCopy);
+
+    // si la playlist courante est celle des favoris, alors on l'actualise
+    if (currentPlaylist?.id === FAVORITES_PLAYLIST_ID) {
+      setCurrentPlaylist(playlistCopy);
+      // si le favori qu'on vient d'enlever est la track jouée, alors on actualise l'index de la track courante
+      if (currentTrackIndex === trackIndex) {
+        setCurrentTrackIndex(trackIndex - 1);
+      }
     }
 
-    const favoritesObjCopy = { ...favoritesObj };
-    delete favoritesObjCopy[id];
-    setFavoritesObj(favoritesObjCopy);
+    // on enlève le favori du local storage
+    removeFavorite(id);
   };
   //#endregion