+++ src/contexts/PlayerContext.tsx
@@ -1,13 +1,21 @@
-import React from 'react';
-import PlaylistTrack from '../models/playlistTrack.model';
+import { createContext, Dispatch, SetStateAction } from 'react';
+import Playlist from '../models/playlist.model';
 
 interface PlayerContextState {
-  play: (playlistTrack: PlaylistTrack) => void;
+  currentPlaylist?: Playlist;
+  currentTrackIndex: number;
+  play: (playlist: Playlist, index: number) => void;
+  setCurrentTrackIndex: Dispatch<SetStateAction<number>>;
+  setCurrentPlaylist: Dispatch<SetStateAction<Playlist | undefined>>;
 }
 
 const defaultState = {
+  currentPlaylist: undefined,
+  currentTrackIndex: -1,
   play: () => {},
+  setCurrentTrackIndex: () => {},
+  setCurrentPlaylist: () => {},
 };
 
-const PlayerContext = React.createContext<PlayerContextState>(defaultState);
+const PlayerContext = createContext<PlayerContextState>(defaultState);
 export default PlayerContext;