+++ src/pages/RootPage/RootPage.tsx
@@ -2,28 +2,49 @@ import { FC } from 'react';
 import { Outlet } from 'react-router-dom';
 import Player from '../../components/Player/Player';
 import Sidebar from '../../components/Sidebar/Sidebar';
+import PlayerContext from '../../contexts/PlayerContext';
 import PlaylistsPreviewsContext from '../../contexts/PlaylistsPreviewsContext';
 import './RootPage.css';
 import useRoot from './useRoot';
 
 const RootPage: FC = () => {
   //#region [logic]
-  const { error, loading, playlistsPreviews } = useRoot();
+  const {
+    error,
+    loading,
+    playlistsPreviews,
+    currentPlaylist,
+    currentTrackIndex,
+    currentTrack,
+    play,
+    setCurrentTrackIndex,
+    setCurrentPlaylist,
+  } = useRoot();
   //#endregion
 
   //#region [render]
   return (
-    <PlaylistsPreviewsContext.Provider
-      value={{ error, loading, playlistsPreviews }}
+    <PlayerContext.Provider
+      value={{
+        currentPlaylist,
+        currentTrackIndex,
+        play,
+        setCurrentTrackIndex,
+        setCurrentPlaylist,
+      }}
     >
-      <div className="root-page">
-        <div className="root-top-container">
-          <Sidebar className="root-sidebar" />
-          <Outlet />
+      <PlaylistsPreviewsContext.Provider
+        value={{ error, loading, playlistsPreviews }}
+      >
+        <div className="root-page">
+          <div className="root-top-container">
+            <Sidebar className="root-sidebar" />
+            <Outlet />
+          </div>
+          <Player track={currentTrack} />
         </div>
-        <Player />
-      </div>
-    </PlaylistsPreviewsContext.Provider>
+      </PlaylistsPreviewsContext.Provider>
+    </PlayerContext.Provider>
   );
   //#endregion
 };